+++ tests/examples/script_single_optional/main.sh
@@ -1,0 +1,41 @@
+#!/usr/bin/env -S bash -x
+# Exit on first error and clean up.
+set -e
+trap 'kill $(pgrep -g $$ | grep -v $$) > /dev/null 2> /dev/null || :' EXIT
+rm -rvf $(cat .gitignore)
+
+# Run the example with the first plan1.py
+cp plan1.py plan.py
+stepup boot -n 1 -w & # > current_stdout1.txt &
+
+# Get the graph after completion of the pending steps.
+stepup wait
+stepup graph current_graph1
+stepup join
+
+# Wait for background processes, if any.
+wait
+
+# Check files that are expected to be present and/or missing.
+[[ -f plan.py ]] || exit 1
+[[ -f generate.py ]] || exit 1
+[[ ! -f runout.txt ]] || exit 1
+[[ ! -f copy.txt ]] || exit 1
+
+# Run the example with the second plan2.py
+cp plan2.py plan.py
+stepup boot -n 1 -w & # > current_stdout2.txt &
+
+# Get the graph after completion of the pending steps.
+stepup wait
+stepup graph current_graph2
+stepup join
+
+# Wait for background processes, if any.
+wait
+
+# Check files that are expected to be present and/or missing.
+[[ -f plan.py ]] || exit 1
+[[ -f generate.py ]] || exit 1
+[[ -f runout.txt ]] || exit 1
+[[ -f copy.txt ]] || exit 1