+++ tests/examples/script_single_optional/plan1.py
@@ -1,0 +1,5 @@
+#!/usr/bin/env python3
+from stepup.core.api import script, static
+
+static("generate.py")
+script("generate.py", optional=True)