+++ tests/examples/script_single_optional/plan2.py
@@ -1,0 +1,6 @@
+#!/usr/bin/env python3
+from stepup.core.api import copy, script, static
+
+static("generate.py")
+script("generate.py", optional=True)
+copy("runout.txt", "copy.txt")